Development

+++ src/main/java/com/rendyrobbani/keuangan/core/domain/entity/master/classification/kelompok/DataMasterKelompok.java
@@ -53,7 +53,7 @@ public interface DataMasterKelompok extends DataMasterClassificationEntity {
 	@Override
 	String deletedBy();
 
-	Integer level();
+	Byte level();
 
 	String kelompok1Id();
