+++ src/main/java/com/rendyrobbani/keuangan/core/domain/entity/master/classification/sumber/DataMasterSumber.java
@@ -50,7 +50,7 @@ public interface DataMasterSumber extends DataMasterClassificationEntity {
 	@Override
 	String deletedBy();
 
-	Integer level();
+	Byte level();
 
 	String sumber1Id();
